+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: lballiot <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/14 10:22:49 by lballiot          #+#    #+#             */
-/*   Updated: 2017/12/11 13:58:23 by lballiot         ###   ########.fr       */
+/*   Updated: 2017/12/14 12:44:01 by lballiot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ char	*ft_strnstr(const char *str, const char *to_find, size_t n)
 	while (*str && n)
 	{
 		if (ft_memcmp(str, to_find, ft_strlen(to_find)) == 0
-				&& (size_t)n >= ft_strlen(to_find))
+				&& n >= ft_strlen(to_find))
 			return ((char *)str);
 		str++;
 		n--;
